feat: update size max file to 20MB and modify logic notification telegram

+++ database/migrations/2025_06_22_090311_add_certificate_status_to_users_table.php
@@ -14,7 +14,9 @@ return new class extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->enum('certificate_status', ['taken', 'not_taken'])->default('not_taken')->after('exam_status');
+            if (!Schema::hasColumn('users', 'certificate_status')) {
+                $table->enum('certificate_status', ['taken', 'not_taken'])->default('not_taken')->after('exam_status');
+            }
         });
     }
 
@@ -26,7 +28,9 @@ return new class extends Migration
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('certificate_status');
+            if (Schema::hasColumn('users', 'certificate_status')) {
+                $table->dropColumn('certificate_status');
+            }
         });
     }
 };
